Background applies correctly now.
It's necessary to apply the background filter early
in the URL chain. If you define a URL like:

    /resize/240x>/background/red/

then the picture will result with a red background
but it will be the original size. I'm not familiar enough
with ImageMagick to understand why at this point, so
just put your background filter at the beginning.

    /background/grey/resize/240x>/

This will work.